Update beta folder

+++ beta/dbconn.py
@@ -15,7 +15,9 @@ class CgBase:
         return MySQLdb.connect(host=dbdetails.host,
                              user=dbdetails.user,
                              passwd=dbdetails.passwd,
-                             db=dbdetails.db)
+                             db=dbdetails.db,
+                             use_unicode=True,
+                             charset='utf8')
     def close_connection(self):
         self.db.close()
 
@@ -146,41 +148,61 @@ class CgBase:
             self.db.rollback()
         return success
 
-    def delete_purchase(self, syncId):
+    def delete_purchase(self, syncId, delete_cart=False):
         syncStr = str(syncId)
         try:
             self.cur.execute("DELETE FROM purchases WHERE syncId = " + syncStr)
-            self.cur.execute("DELETE FROM cart WHERE syncId = " + syncStr)
+            if delete_cart:
+                self.cur.execute("DELETE FROM cart WHERE syncId = " + syncStr)
             self.db.commit()
         except Exception as e:
             self.db.rollback()
             print "Someting weird happened: ", e
+            return False
+        return True
+
+    def delete_cart(self, syncId, boxId):
+        syncStr = str(syncId)
+        boxStr = str(boxId)
+        try:
+            self.cur.execute("DELETE FROM cart WHERE syncId = " + syncStr + " AND boxId = " + boxStr)
+            self.db.commit()
+        except Exception as e:
+            self.db.rollback()
+            print "Someting weird happened: ", e
+            return False
         return True
 
     def sync_cart(self, syncId, status, boxId, quantity, price):
+        result = self.fetchone("cart", ["syncId"], "WHERE syncId="+str(syncId) + " AND boxId=" + str(boxId))
         if status == 0:
-            result = self.fetchone("cart", ["syncId"], "WHERE syncId="+str(syncId) + " AND boxId=" + str(boxId))
             if result is None:
                 self.insert_cart(syncId, boxId, quantity, price)
             return True
         elif status == 1: # edited entry
             pass
         elif status == 2: # deleted entry 
-            pass
+            if result is not None:
+                return self.delete_cart(syncId, boxId)
+            else:
+                return True
         elif status == 3: # synced entry 
             pass
         return False
 
     def sync_purchase(self, syncId, status, country, card, discount, date):
+        result = self.fetchone("purchases", ["status"], "WHERE syncId="+str(syncId))
         if status == 0: # new entry
-            result = self.fetchone("purchases", ["status"], "WHERE syncId="+str(syncId))
             if result is None:
                 self.insert_purchase(country, card, date, discount, {}, syncId=syncId)
             return True
         elif status == 1: # edited entry
             pass
         elif status == 2: # deleted entry 
-            pass
+            if result is not None:
+                return self.delete_purchase(syncId)
+            else:
+                return True
         elif status == 3: # synced entry 
             pass
         return False
@@ -190,3 +212,29 @@ class CgBase:
             self.update("purchases", {"status": 3}, True, "WHERE syncId="+str(syncId))
         else:
             self.update("cart", {"status": 3}, True, "WHERE syncId="+str(syncId)+" AND boxId=" + str(boxId))
+
+    def get_boxes(self):
+        boxes = OrderedDict()
+        results = self.fetchall("boxes", ["boxesEntryId", "title", "price"])
+        for result in results:
+            (boxId, title, price) = result
+            boxes[boxId] = title
+        return boxes
+
+    def get_box_stats(self):
+        stats = {} 
+        results = self.fetchall("cart, purchases", 
+                                ["cart.boxId", "cart.quantity", "purchases.date"],
+                                "WHERE cart.syncId = purchases.syncId AND cart.status != 2 ORDER BY date ASC")
+        for (boxId, quantity, date) in results:
+            date = date.strftime('%Y-%m-%d')
+            try:
+                foo = stats[date]
+            except:
+                stats[date] = {}
+            try:
+                foo = stats[date][boxId] 
+                stats[date][boxId] += quantity 
+            except:
+                stats[date][boxId] = quantity 
+        return stats
